requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.1
MarkupSafe==2.1.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.1
setuptools==69.0.3
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
wheel==0.38.4

jose~=1.0.0